Modified xpublish to be able to serve multiple datasets
